Refactored `combine` as a free function
